eval/hybrid_search/requirements.txt
wikipedia == 1.4.0
tiktoken == 0.7.0
langchain == 0.3.0
openai == 1.55.3
azure-search-documents == 11.6.0b2
python-dotenv == 1.0.1